Conflicts:
	Lint/marte_components_flint_files.lnt
	Source/Components/GAMs/Makefile.inc
	Test/Components/GAMs/Makefile.cov
	Test/Components/GAMs/Makefile.inc

+++ Source/Components/DataSources/NI6368/NI6368DAC.cpp
@@ -934,7 +934,7 @@ bool NI6368DAC::Synchronise() {
     for (i = 0u; (i < NI6368DAC_MAX_CHANNELS) && (ok); i++) {
         if (dacEnabled[i]) {
             size_t samplesToWrite = numberOfElements[i];
-            uint32 k = 0u;
+            ssize_t k = 0;
             if (channelsMemory[i] != NULL_PTR(float32 *)) {
                 while ((samplesToWrite > 0u) && (ok)) {
                     ssize_t samplesWritten = xseries_write_ao(channelsFileDescriptors[i], &channelsMemory[i][k], samplesToWrite);
